2.42

+++ EInvoice/src/main/webapp/WEB-INF/jsp/userForm.jsp
@@ -1,6 +1,7 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <%@ taglib prefix="form" uri="http://www.springframework.org/tags/form"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
 <%@page session="true"%>
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
     "http://www.w3.org/TR/html4/loose.dtd">
@@ -13,6 +14,7 @@
 <jsp:include page="_menu.jsp"/>
     <div align="center">
         <h1>Create/Edit User</h1>
+        <c:out value = "${message}"/><br>
         <form:form action="saveUser" method="post" modelAttribute="user">
         <table>
             <form:hidden path="id"/>
@@ -32,11 +34,17 @@
             	<td><form:input path="expenseLimit"/></td>
             </tr>
             <tr>
-                <td colspan="2" align="center"><input type="submit" value="Save"></td>
+                <td align="center"><input type="submit" value="Save"></td>
             </tr>
         </table>
         </form:form>
-        <a href="${pageContext.request.contextPath}/user/deleteUser">Delete this user</a>
+        <button onclick="cancel()">Cancel</button>
+        <a href="${pageContext.request.contextPath}/user/deleteUser" onclick="return confirm('Are you sure?')">Delete this user</a>
     </div>
+ <script>
+ 	function cancel() {
+		window.location.href = "${pageContext.request.contextPath}/invoice";
+	}
+ </script>
 </body>
 </html>
